WIP for storage abstraction.  Likely not buildable
