+++ storage/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@nimbella/storage-gcs",
+  "name": "@nimbella/storage-provider",
   "version": "0.0.1",
-  "description": "The google cloud storage provider for nimbella-cli",
+  "description": "The types that must be implemented by every Nimbella storage provider",
   "main": "lib/index.js",
   "repository": {
     "type": "git",
@@ -10,11 +10,7 @@
   "keywords": [],
   "author": "Nimbella Corporation",
   "license": "Apache-2.0",
-  "dependencies": {
-    "@google-cloud/storage": "5.3.0"
-  },
   "devDependencies": {
-    "@nimbella/storage-provider": "0.0.1",
     "rimraf": "^3.0.1",
     "typescript": "^3.4.3"
   },
