+++ storage/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    "importHelpers": true,
+    "module": "commonjs",
+    "outDir": "lib",
+    "rootDir": "src",
+    "target": "es2017"
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
